+++ ACP Monitor.lua
@@ -9,42 +9,51 @@
 -- Resolve Repository Root
 --------------------------------------------------------------------------------
 
-local source = debug.getinfo(1, "S").source
-local scriptPath = source:sub(2):match("(.*/)")
-local repositoryRoot = scriptPath:gsub("/$", "")
+local source =
+    debug.getinfo(1, "S").source
+
+local scriptPath =
+    source:sub(2):match("(.*/)")
+
+local repositoryRoot =
+    scriptPath:gsub("/$", "")
 
 --------------------------------------------------------------------------------
 -- Configure Package Path
 --------------------------------------------------------------------------------
 
-local modulePaths = {
-    repositoryRoot .. "/?.lua",
-    repositoryRoot .. "/?/init.lua"
-}
-
-for _, modulePath in ipairs(modulePaths) do
-    if not package.path:find(modulePath, 1, true) then
-        package.path = package.path .. ";" .. modulePath
-    end
-end
+package.path =
+    package.path
+    .. ";"
+    .. repositoryRoot
+    .. "/?.lua"
+    .. ";"
+    .. repositoryRoot
+    .. "/?/init.lua"
 
 --------------------------------------------------------------------------------
 -- Bootstrap
 --------------------------------------------------------------------------------
 
 local Bootstrap =
-    require("Core.Application.Bootstrap")
+    require(
+        "Core.Application.ApplicationBootstrapper"
+    )
 
 assert(
     Bootstrap.Initialize(),
-    "Unable to initialize ACP Monitor.")
+    "Unable to initialize ACP Monitor."
+)
 
 --------------------------------------------------------------------------------
 -- Application
 --------------------------------------------------------------------------------
 
 local MonitorApplication =
-    require("Core.Application.Monitor.MonitorApplication")
+    require(
+        "Core.Application.Monitor.MonitorApplication"
+    )
 
 MonitorApplication.Initialize()
+
 MonitorApplication.Run()